Android patch: Deduplicate data finding code in tests.
The code to find the data and testdata directories was needlessly
duplicated across cintltst and intltest, causing pain for porting to
other platforms. Extract all this functionality into ctestfw so it
can be shared across the test suites.

The upstream request is tracked at
https://unicode-org.atlassian.net/browse/ICU-13849.

The change was modified in ICU 70. The patch for ICU 69 is
https://r.android.com/1802370

In ICU 74, ctest_dataSrcDir() and ctest_dataOutDir() are removed
from cintltst.c to avoid the duplicated symbols.
https://r.android.com/3005253

Bug: 120776993
Test: treehugger
Change-Id: I609079ab8eecd262e398cd1814d5f52268be4391