Android patch: Allow duplicated fields in DateTimePatternGenerator
The original patch can be found at
https://r.android.com/1458480

A skeleton, e.g. "jmma", is mapped into "ahmma" internally
by #mapSkeletonMetacharacters, and thus the internal skeleton
has duplicated fields "a".

This is a regression when we moved the implementation of
android.text.format.DateFormat#getBestPattern() from
udatpg_getBestPattern in ICU4C to ICU4j DateTimePatternGenerator
in the CL https://r.android.com/1355735.
The expected pattern generated from skeleton of "jmma" is
"h:mm a".

Bug: 170233598
Test: n/a
Change-Id: I56b50d466c9a3c0a95c8ecc41c98f75a8f4bb2b9