DO NOT MERGE Track tzdb 2023d update. [U CTS]
TimeZoneTest#TestCanonicalID is marked as ignored (both icu4j
and icu4c versions) as it depends on tzdata and that makes it bad
CTS test. Test is already ignored on older CTS branches.

TestLocalValidity: ignoring test case added on aosp/main as it
will fail with older tzdata-s. Not ignoring it completely as it
is easy to keep the test  working regardless of the installed
tzdata.

Bug: 323820851
Test: run cts -m CtsIcuTestCases -t android.icu.dev.test.format.TimeZoneFormatTest
Test: and
Test: run cts -m CtsIcuTestCases -t android.icu.dev.test.timezone.TimeZoneTest
Test: on ab/11442300 (2023a) instance.

Change-Id: Iaa6859d792626a0f5b3ce1b92521b380421c52ab
Merged-In: Iaa6859d792626a0f5b3ce1b92521b380421c52ab