DO NOT MERGE Track tzdb 2023d update. [S CTS]
TimeZoneTest#TestCanonicalID was already marked as ignored,
only comment was added / updated.

TestLocalValidity: ignoring test case added on aosp/main as it
will fail with older tzdata-s. Not ignoring it completely as it
is easy to keep the test  working regardless of the installed
tzdata.

Bug: 324052859
Test: run cts -m CtsIcuTestCases -t android.icu.dev.test.format.TimeZoneFormatTest
Test: and
Test: run cts -m CtsIcuTestCases -t android.icu.dev.test.timezone.TimeZoneTest
Test: on ab/11461191 (2022a) instance.

Change-Id: Iaa6859d792626a0f5b3ce1b92521b380421c52ab
Merged-In: Iaa6859d792626a0f5b3ce1b92521b380421c52ab