Add java_aconfig_library as libs in i18n.module.public.api
This change adds the java_aconfig_library module "icu_aconfig_flags_lib"
as a dependency of the java_sdk_library module "i18n.module.public.api".

Therefore, the flag symbols in the current.txt files are replaced to the
literal strings.

This change also adds the aconfig_declaration module in the
"i18n.module.public.api" definition, so that the state of the flag is
properly propagated to metalava.

Test: m update-api && m i18n.module.public.api.stubs.source && inspect exportable stubs output and verify that flagged apis are included
Bug: 324254328
Change-Id: Ib6b46fb12680d1dcd36c9eede8f8014e5bf4b4b4