Update updatecldrdata.py due to ICU 74
Update the steps according to https://unicode-org.github.io/icu/processes/cldr-icu.html
Not all steps are added, as some steps and variables, e.g. ICU_DATA_VER,
seem to be specific to integrate new CLDR version with X - 1 ICU
version, and I think they are not needed in AOSP because we take
the new ICU version pre-integrated with that CLDR version in the
upstream.

Test: tools/updatecldrdata.py
Change-Id: I022b1ebb103cf45d2a95f6364fd2ac50155bbf50