Enable art/tools/ahat to use libicu_jni
`ahat-tests` requires the `libicu_jni.so` at runtime. This change
allows ahat-tests to use them.

This change includes the source code version of `libicu_jni`, while
the prebuilt version of `libicu_jni` is in aosp/3235647.
Since the source code version of `libicu_jni` will later be pushed to
the prebuilt version and used for presubmit tests, it also needs to be
updated synchronously.

Bug: 361494087
Bug: 342330305
Test: m -j libicu_jni; atest ahat-tests
Change-Id: I91a0fbfde0c71997a020556789d8ece7b8a8c7e4