Override api_surface for intra core and	core platform
Preparation for	changes	that add configuration for the API surfaces in
`build/soong/java/metalava/main-config.xml` and pass Metalava the name
of the surface instead of `--show-single-annotation` command line
options. That will avoid duplicating the command line arguments for
each API surface in multiple locations.

While most `java_sdk_library` modules use the correct `api_surface`
that is not true for the `intra-core` and `core-platform` APIs so this
change overrides the default `api_surface` for those.

Bug: 391554590
Test: m checkapi
Change-Id: I2530d1f05236fb817468ee78f60d98dc7f998308